requirements.txt
Django>=3.1.5,<3.2
uWSGI>=2.0.18,<2.1
djangorestframework>=3.12.2,<3.13
psycopg2>=2.8.6,<2.9
django_otp>=1.0.2,<1.1